+++ h2o-algos/src/main/java/hex/svd/SVD.java
@@ -424,7 +424,7 @@ public class SVD extends ModelBuilder<SVDModel,SVDModel.SVDParameters,SVDModel.S
       ++_nobs;
     }
 
-    @Override protected void chunkDone(long n){
+    @Override protected void chunkDone(double n){
       double r = 1.0/_nobs;
       _gram.mul(r);
     }
